Renamed app.js to server.js
